requirements.txt
protobuf==6.31.1
python-dotenv==1.1.0
rich==14.0.0
